pyproject.toml
[tool.poetry]
name = "github-activity-tracker"
version = "0.1.0"
description = ""
authors = ["timok19 <tima1199@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
tortoise-orm = {extras = ["asyncmy"], version = "^0.21.3"}
pydantic = "^2.7.3"
black = "^24.4.2"
pygithub = "^2.3.0"
loguru = "^0.7.2"
fastapi-injector = "^0.6.0"
fastapi-controllers = "^0.3.0"
python-dotenv = "^1.0.1"
mypy = "^1.10.0"
varname = "^0.13.1"
httpx = "^0.27.0"
asgi-lifespan = "^2.1.0"
async-lru = "^2.0.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--durations=0 --verbose --color=yes --showlocals"
python_files = ["tests/**/*_test_case.py"]
xfail_strict = true
log_cli = true
log_cli_level = "INFO"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::RuntimeWarning"
]

[tool.black]
line-length = 150